As the might Lord Vader once said: IT'S WORKING, IT'S WOOOOORKING!

+++ Model/CityBikes.Data/flow-time/flow-time-prediction-matrix.hpp
@@ -5,6 +5,6 @@
 namespace CityBikes::Data::FlowTime
 {
 	template<size_t Nodes>
-	struct FlowTimePredictionMatrix : public std::array<std::array<float, Nodes>, Nodes>
+	struct FlowTimePredictionMatrix : public std::array<std::array<double, Nodes>, Nodes>
 	{ };
 }
