+++ Model/CityBikes.DataProcessing/time/time-processing.hpp
@@ -19,7 +19,7 @@ namespace CityBikes::DataProcessing::Time
 			return (double)timeFrames * (time.tm_hour * 60 + time.tm_min) / 1440;
 		}
 
-		float quantize(size_t seconds) const
+		float quantize(double seconds) const
 		{
 			return (double)timeFrames * seconds / 86400;
 		}
